[CS-3806]:Migrate useAccountSettings hook to ts
